feat(challenge): added UnaunthorizedException for 401 error management
