+++ CHANGELOG.md
@@ -3,6 +3,16 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### [itachallenge-challenge-3.2.2-RELEASE] - 2026-02-12
+
+### Added
+- New com.itachallenge.challenge.exception.UnauthorizedException to represent authentication/token failures in the Challenge API. (GitHub Task [#132], Subtask [#150], PR [#1092])
+- Added global handling for UnauthorizedException to return 401 Unauthorized with a consistent message DTO payload. (GitHub Task [#132], Subtask [#150], PR [#1092]
+
+### Changed
+- Updated ChallengeController to convert token parsing/validation failures into UnauthorizedException so the API returns 401 instead of 500. (GitHub Task [#132], Subtask [#150], PR [#1092])
+  * DELETE /challenges/{id} → 401 Unauthorized when the Authorization header/token is missing, invalid, or expired
+
 ### [itachallenge-user-3.2.1-RELEASE] - 2026-02-10
 
 ### Changed
