+++ itachallenge-challenge/src/main/java/com/itachallenge/challenge/exception/UnauthorizedException.java
@@ -1,0 +1,7 @@
+package com.itachallenge.challenge.exception;
+
+public class UnauthorizedException extends RuntimeException {
+    public UnauthorizedException(String message) {
+        super(message);
+    }
+}
