+++ itachallenge-challenge/src/main/java/com/itachallenge/common/exception/GlobalExceptionHandler.java
@@ -192,4 +192,10 @@ public class GlobalExceptionHandler {
         return ResponseEntity.status(HttpStatus.CONFLICT).body(new MessageDto(ex.getMessage()));
     }
 
+    @ExceptionHandler(UnauthorizedException.class)
+    public ResponseEntity<MessageDto> handleUnauthorized(UnauthorizedException ex) {
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+                .body(new MessageDto(ex.getMessage()));
+    }
+
 }
