+++ itachallenge-challenge/src/test/java/com/itachallenge/challenge/controller/ChallengeControllerErrorTest.java
@@ -2,6 +2,7 @@ package com.itachallenge.challenge.controller;
 
 import com.itachallenge.challenge.dto.MessageDto;
 import com.itachallenge.challenge.exception.ChallengeNotFoundException;
+import com.itachallenge.challenge.exception.UnauthorizedException;
 import com.itachallenge.challenge.service.IChallengeService;
 import com.itachallenge.challenge.service.IChallengeJwtFacade;
 import org.junit.jupiter.api.Assertions;
@@ -51,18 +52,24 @@ class ChallengeControllerErrorTest {
     @Test
     void deleteOneChallenge_notFound() {
         String id = "non_existing_id";
+        String authHeader = "Bearer valid.token";
+        String userId = "existing_userId";
 
+        when(challengeJwtFacade.getUserUuIdFromAuthenticationHeader(authHeader)).thenReturn(userId);
         when(challengeService.deleteChallengeById(id))
-                .thenReturn(Mono.error(
-                        new ChallengeNotFoundException(String.format("Challenge with id: %s not found", id))));
+                .thenReturn(Mono.error(new ChallengeNotFoundException(
+                        String.format("Challenge with id: %s not found", id)
+                )));
 
         webTestClient.delete()
                 .uri("/itachallenge/api/v1/challenge/challenges/" + id)
+                .header("Authorization", authHeader)
                 .exchange()
                 .expectStatus().isNotFound()
                 .expectBody()
-                .jsonPath("$.message")
-                .isEqualTo("Challenge with id: non_existing_id not found");
+                .jsonPath("$.message").isEqualTo("Challenge with id: non_existing_id not found");
+
+        verify(challengeService, times(1)).deleteChallengeById(id);
     }
 
     @Test
@@ -107,5 +114,19 @@ class ChallengeControllerErrorTest {
                 .value(messageDto -> Assertions.assertEquals(errorMessage, messageDto.getMessage()));
 
         verify(challengeService, times(1)).removeChallengeFromBookmarks(challengeId, userId);
+    }
+
+    @Test
+    void deleteOneChallenge_missingHeader_shouldReturn401() {
+        String id = "non_existing_id";
+
+        when(challengeJwtFacade.getUserUuIdFromAuthenticationHeader(null))
+                .thenThrow(new UnauthorizedException("Missing or invalid authorization token"));
+        webTestClient.delete()
+                .uri("/itachallenge/api/v1/challenge/challenges/" + id)
+                .exchange()
+                .expectStatus().isUnauthorized();
+
+        verify(challengeService, never()).deleteChallengeById(anyString());
     }
 }
